Move stopper to utils
